Unexpected Token
I randomly ran into this error again, I had literally fixed all errors,
was writing to localStorage and then this hit. I imagine it's an old
bad habit resurfacing again; hopefully I can fix this prior to due time
or that it's a simple early closing of a function or something?
HOPEFULLY I'm not graded too harshly on this and can learn from the
mistake and redo it 100 times to remember not to do it again!