Fixed Syntax Errors
Fixed all of the errors, feel kind of proud of myself because it was
always a weakness. Now I must fix saving to localStorage as it broke in
the meantime but everything seems to be ok for now and fully working
minus the lack of saving to localStorage. Everything else should turn
out fine and fully working.